Adjustments on Tilemap, Player Inspector and Colliders
Changed the ability bool so the player cannot double jump at the beginning of the game.
Made all values private in the inspector to avoid unintended changes.
Adjusted some of the tiles so it would be much easier to get around in the Tutorial room.
Adjusted the player collider so it fits with the newest version of the character asset.