Animation Controller
Added animation controller for the player to switch between idle, run and jump.
Added jump animation sequence for jump start, jump rise, jump fall and jump land.
Added an animation debugger for checking the animation if the sequence is changing correctly.
Added Composite Collider 2D to the tile map so that the levels are now having a whole collider and the player won't fall into the cracks between tiles.
Modified the player collider offset so it doesn't looks like floating in the air.
Applied alpha is transparency to solve the Spine's warning.